create js enviroment
