+++ src/js/index.js
@@ -1,7 +1,4 @@
-const menuBtn = document.getElementById("menuBtn");
-const sideNav = document.getElementById("sideNav");
+// Entry Point
+import main from "./main";
 
-menuBtn.addEventListener("click", (e) => {
-  menuBtn.classList.toggle("active");
-  sideNav.classList.toggle("active");
-});
+main();
